backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
psycopg2
python-dotenv>=1.0.0
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=2.0.0
requests>=2.32.0
python-multipart>=0.0.9
passlib[bcrypt]
PyJWT
fpdf2
joblib